passed $path to input_view.php from controller moved input_view.php js to its own file used the /Lib/misc/gettext.js functions for translations corrected translation string in process_ui.js called the processlist_ui.processInputs() function on succesfull ajax load
