+++ Modules/input/input_controller.php
@@ -14,7 +14,7 @@
 
 function input_controller()
 {
-    global $mysqli, $redis, $user, $session, $route, $feed_settings,$param,$device;
+    global $mysqli, $redis, $user, $session, $route, $feed_settings, $param, $device, $path;
     
     // requires at least read access
     if (!isset($session['read'])) return false;
@@ -117,8 +117,10 @@ function input_controller()
             $route->format = "html";
             textdomain("messages");
             $deviceModuleInstalled = file_exists('Modules/device');
-            // $deviceModuleInstalled = false;
-            return view("Modules/input/Views/input_view.php",array('deviceModule'=>$deviceModuleInstalled));
+            return view("Modules/input/Views/input_view.php", array(
+                'path' => $path, 
+                'deviceModule' => $deviceModuleInstalled
+            ));
         }    
         else if ($device && $route->action == 'schedule') {
             $route->format = "html";
